repaired cognitive_bias_module.py
